+++ pkg/experiment/metastore/metastore_block_store.go
@@ -19,10 +19,14 @@ type metastoreIndexStore struct {
 }
 
 const (
-	partitionBucketName = "partition"
+	partitionBucketName   = "partition"
+	partitionMetaKeyName  = "meta"
+	emptyTenantBucketName = "-"
 )
 
 var partitionBucketNameBytes = []byte(partitionBucketName)
+var partitionMetaKeyNameBytes = []byte(partitionMetaKeyName)
+var emptyTenantBucketNameBytes = []byte(emptyTenantBucketName)
 
 func getPartitionBucket(tx *bbolt.Tx) (*bbolt.Bucket, error) {
 	bkt := tx.Bucket(partitionBucketNameBytes)
@@ -37,7 +41,7 @@ func (m *metastoreIndexStore) ListPartitions() []index.PartitionKey {
 	err := m.db.boltdb.View(func(tx *bbolt.Tx) error {
 		bkt, err := getPartitionBucket(tx)
 		if err != nil {
-			return err
+			return errors.Wrap(err, "root partition bucket missing")
 		}
 		err = bkt.ForEachBucket(func(name []byte) error {
 			partitionKeys = append(partitionKeys, index.PartitionKey(name))
@@ -56,13 +60,13 @@ func (m *metastoreIndexStore) ReadPartitionMeta(key index.PartitionKey) (*index.
 	err := m.db.boltdb.View(func(tx *bbolt.Tx) error {
 		bkt, err := getPartitionBucket(tx)
 		if err != nil {
-			return err
+			return errors.Wrap(err, "root partition bucket missing")
 		}
 		partBkt := bkt.Bucket([]byte(key))
 		if partBkt == nil {
 			return fmt.Errorf("partition meta not found for %s", key)
 		}
-		data := partBkt.Get([]byte("meta"))
+		data := partBkt.Get(partitionMetaKeyNameBytes)
 		dec := gob.NewDecoder(bytes.NewReader(data))
 		err = dec.Decode(&meta)
 		if err != nil {
@@ -81,7 +85,7 @@ func (m *metastoreIndexStore) ListShards(key index.PartitionKey) []uint32 {
 	err := m.db.boltdb.View(func(tx *bbolt.Tx) error {
 		bkt, err := getPartitionBucket(tx)
 		if err != nil {
-			return err
+			return errors.Wrap(err, "root partition bucket missing")
 		}
 		partBkt := bkt.Bucket([]byte(key))
 		if partBkt == nil {
@@ -103,7 +107,7 @@ func (m *metastoreIndexStore) ListTenants(key index.PartitionKey, shard uint32) 
 	err := m.db.boltdb.View(func(tx *bbolt.Tx) error {
 		bkt, err := getPartitionBucket(tx)
 		if err != nil {
-			return err
+			return errors.Wrap(err, "root partition bucket missing")
 		}
 		partBkt := bkt.Bucket([]byte(key))
 		if partBkt == nil {
@@ -135,7 +139,7 @@ func (m *metastoreIndexStore) ListBlocks(key index.PartitionKey, shard uint32, t
 	err := m.db.boltdb.View(func(tx *bbolt.Tx) error {
 		bkt, err := getPartitionBucket(tx)
 		if err != nil {
-			return err
+			return errors.Wrap(err, "root partition bucket missing")
 		}
 		partBkt := bkt.Bucket([]byte(key))
 		if partBkt == nil {
@@ -170,40 +174,44 @@ func (m *metastoreIndexStore) ListBlocks(key index.PartitionKey, shard uint32, t
 	return blocks
 }
 
-func (m *metastoreIndexStore) LoadBlock(key index.PartitionKey, shard uint32, tenant string, blockId string) *metastorev1.BlockMeta {
-	var block *metastorev1.BlockMeta
-	err := m.db.boltdb.View(func(tx *bbolt.Tx) error {
-		bkt, err := getPartitionBucket(tx)
-		if err != nil {
-			return err
-		}
-		partBkt := bkt.Bucket([]byte(key))
-		if partBkt == nil {
-			return nil
-		}
-		shardBktName := make([]byte, 4)
-		binary.BigEndian.PutUint32(shardBktName, shard)
-		shardBkt := partBkt.Bucket(shardBktName)
-		if shardBkt == nil {
-			return nil
-		}
-		tenantBkt := shardBkt.Bucket([]byte(tenant))
-		if tenantBkt == nil {
-			return nil
-		}
-		blockData := tenantBkt.Get([]byte(blockId))
-		if blockData == nil {
-			return nil
-		}
-		var md metastorev1.BlockMeta
-		if err := md.UnmarshalVT(blockData); err != nil {
-			return fmt.Errorf("failed to unmarshal block %q: %w", blockId, err)
-		}
-		block = &md
-		return nil
-	})
+func updateBlockMetadataBucket(tx *bbolt.Tx, partitionMeta *index.PartitionMeta, shard uint32, tenant string, fn func(*bbolt.Bucket) error) error {
+	bkt, err := getPartitionBucket(tx)
 	if err != nil {
-		panic(err)
+		return errors.Wrap(err, "root partition bucket missing")
 	}
-	return block
+
+	partBkt, err := getOrCreateSubBucket(bkt, []byte(partitionMeta.Key))
+	if err != nil {
+		return errors.Wrapf(err, "error creating partition bucket for %s", partitionMeta.Key)
+	}
+
+	var buf bytes.Buffer
+	enc := gob.NewEncoder(&buf)
+	err = enc.Encode(partitionMeta)
+	if err != nil {
+		return errors.Wrapf(err, "could not encode partition meta for %s", partitionMeta.Key)
+	}
+
+	err = partBkt.Put(partitionMetaKeyNameBytes, buf.Bytes())
+	if err != nil {
+		return errors.Wrapf(err, "could not write partition meta for %s", partitionMeta.Key)
+	}
+
+	shardBktName := make([]byte, 4)
+	binary.BigEndian.PutUint32(shardBktName, shard)
+	shardBkt, err := getOrCreateSubBucket(partBkt, shardBktName)
+	if err != nil {
+		return errors.Wrapf(err, "error creating shard bucket for partiton %s and shard %d", partitionMeta.Key, shard)
+	}
+
+	tenantBktName := []byte(tenant)
+	if len(tenantBktName) == 0 {
+		tenantBktName = emptyTenantBucketNameBytes
+	}
+	tenantBkt, err := getOrCreateSubBucket(shardBkt, tenantBktName)
+	if err != nil {
+		return errors.Wrapf(err, "error creating tenant bucket for partition %s, shard %d and tenant %s", partitionMeta.Key, shard, tenant)
+	}
+
+	return fn(tenantBkt)
 }
